05-20-02.

+++ Day05/src/main/webapp/WEB-INF/views/getBoard.jsp
@@ -4,6 +4,12 @@
 "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
+<!-- Latest compiled and minified CSS -->
+<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" integrity="sha384-HSMxcRTRxnN+Bdg0JdbxYKrThecOKuH5zCYotlSAcp1+c8xmyTe9GYg1l9a69psu" crossorigin="anonymous">
+<!-- Optional theme -->
+<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap-theme.min.css" integrity="sha384-6pzBo3FDv/PJ8r2KRkGHifhEocL+1X2rVCTTkUfGk7/0pbek5mMa1upzvWbrUbOZ" crossorigin="anonymous">
+<!-- Latest compiled and minified JavaScript -->
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js" integrity="sha384-aJ21OjlMXNL5UyIl/XNwTMqvzeRMZH2w8c5cRVpzpU8Y5bApTppSuUkhZXN0VxHd" crossorigin="anonymous"></script>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <title>글 상세</title>
 </head>
@@ -89,9 +95,12 @@
 			            +'&searchKeyword='+encodeURIComponent(searchKeyword);
   }
   function go2(seq, pageNum, searchCondition, searchKeyword){
-	 location.href='deleteBoard.do?seq='+seq+'&pageNum='+pageNum
+	 if(confirm("정말삭제할까요?")){
+	  location.href='deleteBoard.do?seq='+seq+'&pageNum='+pageNum
 			           +'&searchCondition='+searchCondition+'&searchKeyword='+encodeURIComponent(searchKeyword);
+	  }
   }
+	 
   function go3(seq, pageNum, searchCondition, searchKeyword,re_ref,re_lev,re_seq){
 		 location.href='replyBoard.do?seq='+seq+'&pageNum='+pageNum
 				           +'&searchCondition='+searchCondition+'&searchKeyword='+encodeURIComponent(searchKeyword)+'&re_ref='+re_ref+'&re_lev='+re_lev+"&re_seq="+re_seq;
